Change getWallet method visibility and add route
The getWallet method in FinancialController was changed from private to public. This was necessary to add a new route in Routing.kt file to get a wallet associated with a specific userId. Now, an HTTP GET request can be made to fetch a wallet's information using the '/getWallet/{userId}' endpoint.