Revise transaction amount type to Double and add method documentation
Changed the 'amount' field in the 'Transaction' data class and methods in 'FinancialController' from 'Int' to 'Double' to allow floating point amounts. Extensive documentation for controller methods was also added, clarifying their usage and potential exceptions. The routes in the 'Routing' file were updated accordingly to handle the amount of Double type.