Change transaction creation from patch to post and improve error message
The code modifies the HTTP method from patch to post when creating a new transaction in the FinancialController. It also improves the exception handling by appending the HTTP response status to the 'Error while creating transaction' exception message. Additionally, it changes the datatype of 'balance' in the Wallet entity from Integer to Double.